Always use np serializer when json dumping
